+++ mooncake-store/tests/offset_allocator_test.cpp
@@ -4,7 +4,7 @@
 #include <random>
 #include <vector>
 
-#include "offset_allocator/offsetAllocator.hpp"
+#include "offset_allocator/offset_allocator.hpp"
 
 using namespace mooncake::offset_allocator;
